+++ lua/obsidian/img_paste.lua
@@ -6,6 +6,15 @@ local util = require "obsidian.util"
 
 local M = {}
 
+local img_types = {
+  ["image/jpeg"] = "jpeg",
+  ["image/png"] = "png",
+  ["image/avif"] = "avif",
+  ["image/webp"] = "webp",
+  ["image/bmp"] = "bmp",
+  ["image/gif"] = "gif",
+}
+
 -- Image pasting adapted from https://github.com/ekickx/clipboard-image.nvim
 
 ---@return string
@@ -29,10 +38,19 @@ local function get_clip_check_command()
   return check_cmd
 end
 
---- Check if clipboard contains image data.
+local function get_image_type(content)
+  for _, line in ipairs(content) do
+    if img_types[line] then
+      return img_types[line]
+    end
+  end
+  return nil
+end
+
+--- Get the type of image on the clipboard.
 ---
----@return boolean
-function M.clipboard_is_img()
+---@return "png"|"jpeg"|nil
+function M.get_clipboard_img_type()
   local check_cmd = get_clip_check_command()
   local result_string = vim.fn.system(check_cmd)
   local content = vim.split(result_string, "\n")
@@ -41,47 +59,56 @@ function M.clipboard_is_img()
   -- See: [Data URI scheme](https://en.wikipedia.org/wiki/Data_URI_scheme)
   local this_os = api.get_os()
   if this_os == api.OSType.Linux or this_os == api.OSType.FreeBSD then
-    if vim.tbl_contains(content, "image/png") then
-      is_img = true
-    elseif vim.tbl_contains(content, "text/uri-list") then
+    if vim.tbl_contains(content, "text/uri-list") then
       local success =
         os.execute "wl-paste --type text/uri-list | sed 's|file://||' | head -n1 | tr -d '[:space:]' | xargs -I{} sh -c 'wl-copy < \"$1\"' _ {}"
-      is_img = success == 0
+      if success == 0 then
+        -- Re-check for image type after potential conversion
+        result_string = vim.fn.system(check_cmd)
+        content = vim.split(result_string, "\n")
+        return get_image_type(content)
+      end
+    else
+      return get_image_type(content)
     end
+
+  -- Code for non-Linux Operating systems (only supports png)
   elseif this_os == api.OSType.Darwin then
     is_img = string.sub(content[1], 1, 9) == "iVBORw0KG" -- Magic png number in base64
+    if is_img then
+      return "png"
+    end
   elseif this_os == api.OSType.Windows or this_os == api.OSType.Wsl then
     is_img = content ~= nil
+    if is_img then
+      return "png"
+    end
   else
     error("image saving not implemented for OS '" .. this_os .. "'")
   end
-  return is_img
+  return nil
 end
 
---- TODO: refactor with run_job?
+--- TODO: refactor Windows with run_job?
 
 --- Save image from clipboard to `path`.
 ---@param path string
+---@param img_type "png" | "jpeg"
 ---
 ---@return boolean|integer|? result
-local function save_clipboard_image(path)
+local function save_clipboard_image(path, img_type)
   local this_os = api.get_os()
 
   if this_os == api.OSType.Linux or this_os == api.OSType.FreeBSD then
+    local mime_type = "image/" .. img_type
     local cmd
     local display_server = os.getenv "XDG_SESSION_TYPE"
     if display_server == "x11" or display_server == "tty" then
-      cmd = string.format("xclip -selection clipboard -t image/png -o > '%s'", path)
-    elseif display_server == "wayland" then
-      cmd = string.format("wl-paste --no-newline --type image/png > %s", vim.fn.shellescape(path))
+      cmd = string.format("xclip -selection clipboard -t %s -o > '%s'", mime_type, path)
       return run_job { "bash", "-c", cmd }
-    end
-
-    local result = os.execute(cmd)
-    if type(result) == "number" and result > 0 then
-      return false
-    else
-      return result
+    elseif display_server == "wayland" then
+      cmd = string.format("wl-paste --no-newline --type %s > %s", mime_type, vim.fn.shellescape(path))
+      return run_job { "bash", "-c", cmd }
     end
   elseif this_os == api.OSType.Windows or this_os == api.OSType.Wsl then
     local cmd = 'powershell.exe -c "'
@@ -96,7 +123,7 @@ local function save_clipboard_image(path)
 end
 
 --- @param path string image_path The absolute path to the image file.
-M.paste = function(path)
+M.paste = function(path, img_type)
   if util.contains_invalid_characters(path) then
     log.warn "Links will not work with file names containing any of these characters in Obsidian: # ^ [ ] |"
   end
@@ -104,12 +131,15 @@ M.paste = function(path)
   ---@diagnostic disable-next-line: cast-local-type
   path = Path.new(path)
 
-  -- Make sure fname ends with ".png"
+  -- If there is no suffix provided, append it
   if not path.suffix then
     ---@diagnostic disable-next-line: cast-local-type
-    path = path:with_suffix ".png"
-  elseif path.suffix ~= ".png" then
-    return log.err("invalid suffix for image name '%s', must be '.png'", path.suffix)
+    path = path:with_suffix("." .. img_type)
+
+  -- If user appends their own suffix, check if it is valid based on img_type
+  elseif not (path.suffix == "." .. img_type or (img_type == "jpeg" and path.suffix == ".jpg")) then
+    local expected_suffix = (img_type == "jpeg") and ".jpeg' or '.jpg" or "." .. img_type
+    return log.err("invalid suffix for image name '%s', must be '%s'", path.suffix, expected_suffix)
   end
 
   if Obsidian.opts.attachments.confirm_img_paste then
@@ -123,7 +153,7 @@ M.paste = function(path)
   assert(path:parent()):mkdir { exist_ok = true, parents = true }
 
   -- Paste image.
-  local result = save_clipboard_image(tostring(path))
+  local result = save_clipboard_image(tostring(path), img_type)
   if result == false then
     log.err "Failed to save image"
     return